chore: CR changes and repetition reduced
